@@ -15,11 +15,11 @@ Gem::Specification.new do |s|
 
   s.files = Dir["{app,config,db,lib}/**/*"] + ["MIT-LICENSE", "Rakefile", "README.rdoc"]
 
-  s.add_dependency "rails", "~> 3.2.11"
+  s.add_dependency "rails", "~> 4.0.0"
   
   # Module dependencies  
   s.add_dependency 'acts_as_markup'
-  s.add_dependency 'acts_as_published', '~> 0.0.8.pre'
+  s.add_dependency 'acts_as_published'
   s.add_dependency 'friendly_id'
   s.add_dependency 'kaminari'
   s.add_dependency 'bootstrap-kaminari-views'
@@ -46,5 +46,5 @@ Gem::Specification.new do |s|
   # s.add_development_dependency 'guard-rails' # See gemfile for a patched version, for engine support.
   s.add_development_dependency 'guard-rspec'
   s.add_development_dependency 'guard-bundler'  
-  s.add_development_dependency 'rb-inotify', '~> 0.8.8'
+  s.add_development_dependency 'rb-inotify'
 end
